% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine}
\alias{cosine}
\title{Calculate cosine similarity between rows of \code{a} and \code{b} (\code{b} could be a vector)}
\usage{
cosine(a, b = NULL)
}
\arguments{
\item{a}{matrix}

\item{b}{matrix or vector, if null, then \code{b} is set to be equal to \code{a}}
}
\value{
matrix of cosine distances between rows of \code{a} and rows of \code{b}
}
\description{
Calculate cosine similarity between rows of \code{a} and \code{b} (\code{b} could be a vector)
}
\keyword{internal}
