% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{getHeatmaps}
\alias{getHeatmaps}
\title{Get files with modules' heatmaps}
\usage{
getHeatmaps(
  modules,
  patterns,
  E,
  E.prep,
  order,
  data.annotation = NULL,
  network.annotation,
  width = 25,
  height = 10,
  work.dir = work.dir
)
}
\arguments{
\item{modules}{Metabolic modules.}

\item{patterns}{Patterns of metabolic modules.}

\item{order}{Order of samples on the heatmap.}

\item{data.annotation}{Vector of sample annotation with the same sample order
as columns in patterns}

\item{network.annotation}{Metabolic network annotation.}

\item{width}{Width of the figure.}

\item{height}{Height of the figure.}

\item{work.dir}{Working directory where results should be saved.}

\item{gene.exprs}{Gene expression.}
}
\value{
Results of this function can be seen in work.dir (module heatmaps as .png files and all patterns common heatmap as m.centers.pdf file).
}
\description{
Get files with modules' heatmaps
}
