% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{getModulesAnnotation}
\alias{getModulesAnnotation}
\title{Get files with modules' annotations}
\usage{
getModulesAnnotation(network.annotation, nets, work.dir, padj.threshold = Inf)
}
\arguments{
\item{network.annotation}{Metabolic network annotation.}

\item{nets}{Scored networks.}

\item{work.dir}{Working directory where files with module genes are (results will be saved here as well).}

\item{padj.threshold}{Threshold, adjusted p-value, for pathways.}
}
\value{
Results of this function can be seen in work.dir (.tsv files).
}
\description{
Get files with modules' annotations
}
