% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{prepareNetwork}
\alias{prepareNetwork}
\title{Prepare network}
\usage{
prepareNetwork(
  E,
  network,
  topology = c("metabolites", "atoms"),
  met.to.filter = data.table::fread(system.file("mets2mask.lst", package =
    "GAMclust"))$ID,
  network.annotation,
  gene2reaction.extra = NULL
)
}
\arguments{
\item{E}{Expression matrix after the \code{prepareData()} function.}

\item{network}{Metabolic network.}

\item{topology}{Vertices can be represented either as \code{metabolites}, either as \code{atoms}.}

\item{met.to.filter}{Metabolites that should not be used as connections in the module.}

\item{network.annotation}{Metabolic network annotation.}

\item{gene2reaction.extra}{For a combined network: supplementary file with genes that either come from proteome or are not linked to a specific enzyme.}
}
\value{
Edges of the final network.
}
\description{
Constructs a metabolic interaction network tailored to the analyzed dataset.
The function integrates network topology with gene–enzyme–reaction
annotations, removes filtered metabolites, and retains only genes present
in the processed expression matrix.
Networks can be represented at either the metabolite or atom level.
The final output corresponds to the largest connected component of the
metabolic network and serves as the graph structure for GAM-clustering.
}
