% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameAtoms}
\alias{renameAtoms}
\title{Rename atoms (to human readable format)}
\usage{
renameAtoms(g, network, metabolites.annotation)
}
\arguments{
\item{g}{Input graph}

\item{network}{Metabolic network with slot `atoms` containing information on
*metabolite* and *element*}

\item{metabolites.annotation}{Metabolites annotation.}
}
\value{
Graph with renamed atoms (original naming can be found as vertices attribute named *atom*)
}
\description{
Rename atoms (to human readable format)
}
