% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_batch_solver}
\alias{seq_batch_solver}
\title{Function for batch solving SGMWCS problems}
\usage{
seq_batch_solver(mwcs_solver)
}
\arguments{
\item{mwcs_solver}{SGMWCS solver from mwcsr package}
}
\description{
Creates a wrapper for sequentially solving multiple SGMWCS optimization
instances. The returned function applies a selected MWCS solver to each
network independently and collects the results into a single list.
}
