% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{gamClustering}
\alias{gamClustering}
\title{GAM-clustering analysis}
\usage{
gamClustering(
  E.prep,
  network.prep,
  cur.centers,
  repeats,
  batch.solver = seq_batch_solver(solver),
  max.module.size = 50,
  start.base = 1,
  base.dec = 0.05,
  show.intermediate.clustering = TRUE,
  verbose = TRUE,
  collect.stats = TRUE,
  work.dir
)
}
\arguments{
\item{E.prep}{Expression matrix after the `prepareData()` function.}

\item{network.prep}{Network edge table driven from `prepareNetwork()` function.}

\item{cur.centers}{Initial patterns produced by `preClustering()` function.}

\item{repeats}{Here you may collapse biological replicas by providung vector with repeated sample names. Otherwise, just use vector with unique sample names.}

\item{batch.solver}{Solver for SGMWCS problem.}

\item{max.module.size}{Maximal number of unique genes in the final module.}

\item{start.base}{The distance to the fake pattern which influences `max.module.size` parameter. Equal to 1 at the beginning.}

\item{base.dec}{The value by which `base` parameter should be reduced if some module's size is bigger that `max.module.size`.}

\item{show.intermediate.clustering}{Whether to show or not heatmap of intermideate clusters.}

\item{verbose}{Verbose running.}

\item{collect.stats}{Whether to save or not running statistics.}

\item{work.dir}{Working directory where results should be saved.}

\item{E}{Expression matrix.}
}
\value{
results$modules -- Metabolic modules.

results$nets -- Scored networks.

results$patterns.pos -- Modules' patterns (genes with positive score only considered).

results$patterns.all -- Modules' patterns (all genes considered).

results$iter.stats -- Statistics from iterations.
}
\description{
GAM-clustering analysis
}
