% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{getAnnotationHeatmap}
\alias{getAnnotationHeatmap}
\title{Heatmap for annotated pathways}
\usage{
getAnnotationHeatmap(
  work.dir,
  padj.threshold = Inf,
  threshold = 0.05,
  file_name = "Modules_heatmap.png"
)
}
\arguments{
\item{work.dir}{Directory with gene and pathways tables}

\item{padj.threshold}{Set threshold for p-adjusted}

\item{threshold}{Minimal percent of genes in the pathway}

\item{file_name}{Name of file with heatmap}
}
\value{
Results of this function can be seen in work.dir (png files)
}
\description{
Heatmap for annotated pathways
}
