% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{getGeneTables}
\alias{getGeneTables}
\title{Get files with modules' genes}
\usage{
getGeneTables(
  modules,
  nets,
  patterns,
  gene.exprs,
  network.annotation,
  work.dir = work.dir
)
}
\arguments{
\item{modules}{Metabolic modules.}

\item{nets}{Scored networks.}

\item{patterns}{Patterns of metabolic modules.}

\item{gene.exprs}{Gene expression.}

\item{network.annotation}{Metabolic network annotation.}

\item{work.dir}{Working directory where results should be saved.}
}
\value{
m.*.genes.tsv -- module genes.

m.*.notInModule.genes.tsv -- genes with positive score not included into module.

m.*.complete.genes.tsv -- top 300 of all genes sorted by correlation value.

Results of this function can be seen in work.dir (three .tsv files for each module with gene lists).
}
\description{
Get files with modules' genes
}
