% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{getGraphs}
\alias{getGraphs}
\title{Get files with modules' graphs}
\usage{
getGraphs(
  modules,
  network.annotation,
  metabolites.annotation,
  rename.atoms = FALSE,
  seed.for.layout = 22,
  work.dir = work.dir
)
}
\arguments{
\item{modules}{Metabolic modules.}

\item{network.annotation}{Metabolic network annotation.}

\item{metabolites.annotation}{Metabolites annotation.}

\item{rename.atoms}{Logical, if using atom based metabolic network rename}

\item{seed.for.layout}{Random seed required for gatom to draw metabolic graph.
each atom to corresponding metabolite's name}

\item{work.dir}{Working directory where results should be saved.}
}
\value{
Results of this function can be seen in work.dir (.pdf, .png and .xgmml files).
}
\description{
Get files with modules' graphs
}
