% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{preClustering}
\alias{preClustering}
\title{Defining initial patterns}
\usage{
preClustering(
  E.prep,
  network.prep,
  initial.number.of.clusters = 32,
  network.annotation,
  use.ICA = FALSE
)
}
\arguments{
\item{E.prep}{Expression matrix after the \code{prepareData()} function.}

\item{network.prep}{Network edge table driven from \code{prepareNetwork()} function.}

\item{initial.number.of.clusters}{The number of clusters for the initial approximation of modules.}

\item{network.annotation}{Metabolic network annotation.}
}
\value{
Initial patterns.
}
\description{
Generates an initial set of transcriptional patterns that serve as
starting points for module detection. The function first restricts the
expression matrix to metabolic genes represented in the network and then
performs clustering to identify preliminary expression signatures.
By default, cluster centers are obtained using k-means clustering.
These initial patterns are subsequently refined by the GAM-clustering algorithm.
}
