% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{preClustering}
\alias{preClustering}
\title{Defining initial patterns}
\usage{
preClustering(
  E.prep,
  network.prep,
  initial.number.of.clusters = 32,
  network.annotation,
  use.ICA = FALSE
)
}
\arguments{
\item{E.prep}{Expression matrix after the `prepareData()` function.}

\item{network.prep}{Network edge table driven from `prepareNetwork()` function.}

\item{initial.number.of.clusters}{The number of clusters for the initial approximation of modules.}

\item{network.annotation}{Metabolic network annotation.}
}
\value{
Initial patterns.
}
\description{
Defining initial patterns
}
