% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{prepareData}
\alias{prepareData}
\title{Prepare gene expression matrix}
\usage{
prepareData(E, use.PCA = TRUE, keep.12k.genes = TRUE, network.annotation)
}
\arguments{
\item{E}{Expression matrix with rownames as gene symbols.}

\item{network.annotation}{Metabolic network annotation.}
}
\value{
Expression matrix prepared for the analysis.
}
\description{
Prepare gene expression matrix
}
