% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{prepareNetwork}
\alias{prepareNetwork}
\title{Prepare network}
\usage{
prepareNetwork(
  E,
  network,
  topology = c("atoms", "metabolites"),
  met.to.filter = data.table::fread(system.file("mets2mask.lst", package =
    "GAMclust"))$ID,
  network.annotation
)
}
\arguments{
\item{E}{Expression matrix after the prepareData() function.}

\item{network}{Metabolic network.}

\item{topology}{Vertices can be represented either as `metabolites`, either as `atoms`.}

\item{met.to.filter}{Metabolites that should not be used as connections in the module.}

\item{network.annotation}{Metabolic network annotation.}
}
\value{
Edges of the final network.
}
\description{
Prepare network
}
